Add mock file of lib2mc test
